requirements.txt
# ===============================
# Core Numerical & ML Libraries
# ===============================
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

# ===============================
# Image Processing
# ===============================
pillow==10.2.0
opencv-python==4.8.1.78

# ===============================
# Deep Learning (CPU, Python 3.10)
# ===============================
torch==2.1.2
torchvision==0.16.2

# ===============================
# MLOps Tools
# ===============================
dvc==3.48.4
mlflow==2.10.2

# ===============================
# API & Model Serving
# ===============================
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# ===============================
# Testing (CI/CD)
# ===============================
pytest==7.4.4

# ===============================
# Jupyter Support
# ===============================
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0

# ===============================
# Visualization
# ===============================
matplotlib==3.8.2
seaborn==0.13.1
